requirements.txt
dnspython==2.7.0
et_xmlfile==2.0.0
mysql-connector==2.2.9
openpyxl==3.1.5
pymongo==4.11
python-dotenv==1.0.1
